# Conflicts:
#	lib/public/Model.js
#	lib/public/views/Flps/Overview/index.js
#	lib/public/views/Logs/Overview/index.js
#	lib/public/views/Runs/Overview/index.js
#	test/e2e/logs.test.js

+++ cpp-api-client/src/cpprest-client/model/CreateFlp.h
@@ -74,11 +74,11 @@ public:
     /// <summary>
     /// The unique identifier of this entity.
     /// </summary>
-    int32_t getRunNumber() const;
+    int64_t getRunNumber() const;
     bool runNumberIsSet() const;
     void unsetRunNumber();
 
-    void setRunNumber(int32_t value);
+    void setRunNumber(int64_t value);
 
 
 protected:
@@ -86,7 +86,7 @@ protected:
     bool m_NameIsSet;
     utility::string_t m_Hostname;
     bool m_HostnameIsSet;
-    int32_t m_RunNumber;
+    int64_t m_RunNumber;
     bool m_RunNumberIsSet;
 };
