+++ cpp-api-client/src/cpprest-client/model/Flp.h
@@ -56,47 +56,47 @@ public:
     /// <summary>
     /// Total data volume out from the readout &#39;equipment&#39; component in bytes. Can reach PetaBytes. Updated regularly.
     /// </summary>
-    int32_t getBytesEquipmentReadOut() const;
+    int64_t getBytesEquipmentReadOut() const;
     bool bytesEquipmentReadOutIsSet() const;
     void unsetBytesEquipmentReadOut();
 
-    void setBytesEquipmentReadOut(int32_t value);
+    void setBytesEquipmentReadOut(int64_t value);
 
     /// <summary>
     /// Total data volume out from the readout &#39;fmq&#39; component in bytes. Can reach PetaBytes. Updated regularly.
     /// </summary>
-    int32_t getBytesFairMQReadOut() const;
+    int64_t getBytesFairMQReadOut() const;
     bool bytesFairMQReadOutIsSet() const;
     void unsetBytesFairMQReadOut();
 
-    void setBytesFairMQReadOut(int32_t value);
+    void setBytesFairMQReadOut(int64_t value);
 
     /// <summary>
     /// Total data volume read out in bytes.
     /// </summary>
-    int32_t getBytesProcessed() const;
+    int64_t getBytesProcessed() const;
     bool bytesProcessedIsSet() const;
     void unsetBytesProcessed();
 
-    void setBytesProcessed(int32_t value);
+    void setBytesProcessed(int64_t value);
 
     /// <summary>
     /// Total data volume out from the readout &#39;recording&#39; component in bytes. Can reach PetaBytes. Updated regularly.
     /// </summary>
-    int32_t getBytesRecordingReadOut() const;
+    int64_t getBytesRecordingReadOut() const;
     bool bytesRecordingReadOutIsSet() const;
     void unsetBytesRecordingReadOut();
 
-    void setBytesRecordingReadOut(int32_t value);
+    void setBytesRecordingReadOut(int64_t value);
 
     /// <summary>
     /// Unix timestamp when this entity was created.
     /// </summary>
-    int32_t getCreatedAt() const;
+    int64_t getCreatedAt() const;
     bool createdAtIsSet() const;
     void unsetCreatedAt();
 
-    void setCreatedAt(int32_t value);
+    void setCreatedAt(int64_t value);
 
     /// <summary>
     /// FLP hostname.
@@ -119,11 +119,11 @@ public:
     /// <summary>
     /// Number of subtimeframes processed in this FLP. Updated regularly.
     /// </summary>
-    int32_t getNTimeframes() const;
+    int64_t getNTimeframes() const;
     bool nTimeframesIsSet() const;
     void unsetNTimeframes();
 
-    void setNTimeframes(int32_t value);
+    void setNTimeframes(int64_t value);
 
     /// <summary>
     /// FLP name.
@@ -137,33 +137,33 @@ public:
     /// <summary>
     /// Unix timestamp when this entity was last updated.
     /// </summary>
-    int32_t getUpdatedAt() const;
+    int64_t getUpdatedAt() const;
     bool updatedAtIsSet() const;
     void unsetUpdatedAt();
 
-    void setUpdatedAt(int32_t value);
+    void setUpdatedAt(int64_t value);
 
 
 protected:
-    int32_t m_BytesEquipmentReadOut;
+    int64_t m_BytesEquipmentReadOut;
     bool m_BytesEquipmentReadOutIsSet;
-    int32_t m_BytesFairMQReadOut;
+    int64_t m_BytesFairMQReadOut;
     bool m_BytesFairMQReadOutIsSet;
-    int32_t m_BytesProcessed;
+    int64_t m_BytesProcessed;
     bool m_BytesProcessedIsSet;
-    int32_t m_BytesRecordingReadOut;
+    int64_t m_BytesRecordingReadOut;
     bool m_BytesRecordingReadOutIsSet;
-    int32_t m_CreatedAt;
+    int64_t m_CreatedAt;
     bool m_CreatedAtIsSet;
     utility::string_t m_Hostname;
     bool m_HostnameIsSet;
     int32_t m_Id;
     bool m_IdIsSet;
-    int32_t m_NTimeframes;
+    int64_t m_NTimeframes;
     bool m_NTimeframesIsSet;
     utility::string_t m_Name;
     bool m_NameIsSet;
-    int32_t m_UpdatedAt;
+    int64_t m_UpdatedAt;
     bool m_UpdatedAtIsSet;
 };
